Add (disabled) code to relax calls to PLT stubs
This is useful to have around, but instead of relaxing PLT calls from
auipc/jalr to jal, we relax them to a gp-relative load and a jalr.
This bypasses the PLT entirely (except for the very first call),
reducing dynamic instruction count (including a jump), at the cost of
an extra static instruction per PLT call.

We might revisit this decision.  Sometimes, saving static code size is
more important than BTB pressure or even spatial locality.